refactor(metametrics): extract UserTraitsService from MetaMetricsController
Move trait derivation and profile session refresh out of MetaMetricsController so that work can live independently as part of deprecating the controller.
