+++ app/scripts/controllers/metametrics-controller-method-action-types.ts
@@ -104,11 +104,6 @@ export type MetaMetricsControllerSetMarketingCampaignCookieIdAction = {
   handler: MetaMetricsController['setMarketingCampaignCookieId'];
 };
 
-export type MetaMetricsControllerHandleMetaMaskStateUpdateAction = {
-  type: `MetaMetricsController:handleMetaMaskStateUpdate`;
-  handler: MetaMetricsController['handleMetaMaskStateUpdate'];
-};
-
 export type MetaMetricsControllerTrackTracesAfterMetricsOptInAction = {
   type: `MetaMetricsController:trackTracesAfterMetricsOptIn`;
   handler: MetaMetricsController['trackTracesAfterMetricsOptIn'];
@@ -166,7 +161,6 @@ export type MetaMetricsControllerMethodActions =
   | MetaMetricsControllerSetParticipateInMetaMetricsAction
   | MetaMetricsControllerSetDataCollectionForMarketingAction
   | MetaMetricsControllerSetMarketingCampaignCookieIdAction
-  | MetaMetricsControllerHandleMetaMaskStateUpdateAction
   | MetaMetricsControllerTrackTracesAfterMetricsOptInAction
   | MetaMetricsControllerClearTracesAfterMetricsOptInAction
   | MetaMetricsControllerAddTraceBeforeMetricsOptInAction
