+++ app/scripts/messenger-client-init/messengers/user-traits-service-messenger.test.ts
@@ -1,0 +1,12 @@
+import { Messenger } from '@metamask/messenger';
+import { getRootMessenger } from '../../lib/messenger';
+import { getUserTraitsServiceMessenger } from './user-traits-service-messenger';
+
+describe('getUserTraitsServiceMessenger', () => {
+  it('returns a restricted messenger', () => {
+    const messenger = getRootMessenger();
+    const userTraitsServiceMessenger = getUserTraitsServiceMessenger(messenger);
+
+    expect(userTraitsServiceMessenger).toBeInstanceOf(Messenger);
+  });
+});
